(PC-2062): refacto after review by MEFA
