+++ src/components/layout/CsvTableButton/CsvTableButton.jsx
@@ -2,7 +2,7 @@ import PropTypes from 'prop-types'
 import React, { Component } from 'react'
 
 class CsvTableButton extends Component {
-  redirectToUrl = () => {
+  handleRedirectToUrl = () => {
     const { history, href, location } = this.props
     const { pathname } = location
     const nextUrl = pathname + '/detail'
@@ -14,12 +14,17 @@ class CsvTableButton extends Component {
     return (
       <button
         className="button is-primary is-flex-button"
-        onClick={this.redirectToUrl}
+        onClick={this.handleRedirectToUrl}
+        type="button"
       >
         {children}
       </button>
     )
   }
+}
+
+CsvTableButton.defaultProps = {
+  children: ''
 }
 
 CsvTableButton.propTypes = {
