+++ src/components/layout/CsvTableButton/CsvTableButtonContainer.js
@@ -1,5 +1,4 @@
 import { withRouter } from 'react-router-dom'
 import CsvTableButton from './CsvTableButton'
-import { compose } from 'redux'
 
-export default compose(withRouter)(CsvTableButton)
+export default withRouter(CsvTableButton)
